+++ lib/angarium/configuration.rb
@@ -5,7 +5,7 @@ module Angarium
                   :primary_key_type, :max_response_body_bytes,
                   :auto_disable_endpoint_after, :respect_retry_after,
                   :max_retry_after, :retry_jitter, :signing_secret_grace_period,
-                  :delivery_attempt_retention
+                  :delivery_attempt_retention, :delivering_timeout
 
     def initialize
       @job_queue        = :default
@@ -22,6 +22,7 @@ module Angarium
       @retry_jitter                = 0.15
       @signing_secret_grace_period = 24.hours
       @delivery_attempt_retention  = nil
+      @delivering_timeout          = 15.minutes
     end
   end
 end